feat: 일별 소비 기록에 만족도 뱃지 추가
