+++ apps/web/src/features/expense/model/evaluationLabel.ts
@@ -1,13 +1,6 @@
-import type { EvaluationType } from '@/shared/types/evaluation.types';
-
-const EVALUATION_LABEL_MAP: Record<EvaluationType, string> = {
-  VERY_SATISFIED: '정말 만족했어요',
-  SATISFIED: '만족했어요',
-  NORMAL: '그냥 그랬어요',
-  DISAPPOINTED: '조금 아쉬웠어요',
-  VERY_DISAPPOINTED: '정말 별로였어요',
-};
-
-export const getEvaluationLabel = (evaluationType: EvaluationType): string => {
-  return EVALUATION_LABEL_MAP[evaluationType];
-};
+/**
+ * @deprecated
+ * 실제 구현은 shared/lib/evaluationLabel.ts에 있음.
+ * FSD 규칙 위반이기에 추후 이 파일은 삭제 예정입니다.
+ */
+export { getEvaluationLabel } from '@/shared/lib/evaluationLabel';
